+++ spec/mtime_strategy_spec.rb
@@ -1,3 +1,5 @@
+require_relative "spec_helper_initializer"
+
 describe "Shakapacker::MtimeStrategy" do
   let(:mtime_strategy) { Shakapacker::MtimeStrategy.new }
   let(:manifest_timestamp) { Time.parse("2021-01-01 12:34:56 UTC") }